Simplify precision-specification in tests.
Now that assertEquals considers dtypes in determining tolerance, most
tests don't need explicitly set precision.

Those that do are a few half precision tests on cuda. In this PR, those
are broken out to be handled explicitly, though we may also want to
consider further loosening the tolerance on half-precision.

ghstack-source-id: 5b62f93fed7972842a469fc650da07d85fa2d4c8
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37181